Add InTransactionRetry helper
