+++ internal/handlers/pg/pgdb/pool.go
@@ -24,7 +24,6 @@ import (
 	"github.com/jackc/pgx/v4/pgxpool"
 	"go.uber.org/zap"
 
-	"github.com/FerretDB/FerretDB/internal/util/lazyerrors"
 	"github.com/FerretDB/FerretDB/internal/util/state"
 )
 
@@ -183,41 +182,4 @@ func (pgPool *Pool) SchemaStats(ctx context.Context, schema, collection string) 
 	return &DBStats{
 		Name: schema,
 	}, nil
-}
-
-// InTransaction wraps the given function f in a transaction.
-// If f returns an error, the transaction is rolled back.
-// Errors are wrapped with lazyerrors.Error,
-// so the caller needs to use errors.Is to check the error,
-// for example, errors.Is(err, ErrSchemaNotExist).
-func (pgPool *Pool) InTransaction(ctx context.Context, f func(pgx.Tx) error) (err error) {
-	var tx pgx.Tx
-	if tx, err = pgPool.Begin(ctx); err != nil {
-		err = lazyerrors.Error(err)
-		return
-	}
-
-	defer func() {
-		if err == nil {
-			return
-		}
-		if rerr := tx.Rollback(ctx); rerr != nil {
-			pgPool.Config().ConnConfig.Logger.Log(
-				ctx, pgx.LogLevelError, "failed to perform rollback",
-				map[string]any{"error": rerr},
-			)
-		}
-	}()
-
-	if err = f(tx); err != nil {
-		err = lazyerrors.Error(err)
-		return
-	}
-
-	if err = tx.Commit(ctx); err != nil {
-		err = lazyerrors.Error(err)
-		return
-	}
-
-	return
 }
